added tests
